Successfully created an Include file!

+++ Plasma/AmbushGhostInputDirector.cpp
@@ -37,10 +37,10 @@ void AmbushGhostInputDirector::Update(double delta, Engine::Scene & game)
 		int pointsLeft	= 0;
 		int pointsRight = 0;
 
-		pointsUp	-= Vector2::Distance(positionTile + Vector2( 0,-1), targetTile);
-		pointsDown	-= Vector2::Distance(positionTile + Vector2( 0, 1), targetTile);
-		pointsLeft	-= Vector2::Distance(positionTile + Vector2(-1, 0), targetTile);
-		pointsRight -= Vector2::Distance(positionTile + Vector2( 1, 0), targetTile);
+		pointsUp	-= (int) Vector2::Distance(positionTile + Vector2( 0,-1), targetTile);
+		pointsDown	-= (int)Vector2::Distance(positionTile + Vector2( 0, 1), targetTile);
+		pointsLeft	-= (int)Vector2::Distance(positionTile + Vector2(-1, 0), targetTile);
+		pointsRight -= (int)Vector2::Distance(positionTile + Vector2( 1, 0), targetTile);
 
 		pointsUp	+= canGoUp		? 8 : 0;
 		pointsDown	+= canGoDown	? 8 : 0;
